get_papers_projects/pyproject.toml
[tool.poetry]
name = "get-papers"
version = "0.1.0"
description = "Fetch PubMed papers with non-academic affiliations"
authors = ["Your Name <your_email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.0"
biopython = "^1.79"

[tool.poetry.scripts]
get-papers-list = "get_papers.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"